+++ src/main/java/nu/revitalized/revitalizedwebshop/controllers/SupplementController.java
@@ -49,8 +49,8 @@ public class SupplementController {
             @RequestParam(required = false) Double maxPrice,
             @RequestParam(required = false) Integer minStock,
             @RequestParam(required = false) Integer maxStock,
-            @RequestParam(required = false) Double minRating,
-            @RequestParam(required = false) Double maxRating,
+            @RequestParam(required = false) Integer minRating,
+            @RequestParam(required = false) Integer maxRating,
             @RequestParam(required = false) String contains
     ) {
         List<SupplementDto> dtos = supplementService.getSupplementsByParam(
